remote control error adapted to all screen sizes + code cleanup
