+++ src/layout/MainLayoutPageWrapper/MainLayoutPageWrapper.css
@@ -76,9 +76,7 @@
   .mainLayout__body {
     height: calc(100vh - var(--main-layout-footer-height));
   }
-}
 
-@media (max-width: 600px) {
   .mainLayout__sidebar,
   .mainLayout__body {
     padding-top: var(--main-layout-header-height);
